some fixes to map loading

+++ src/main.rs
@@ -98,7 +98,8 @@ async fn main() {
             id: 0,
         });
 
-        let map = Map::new(vec2(16.0, 16.0),"assets/maps/map_01.json").await;
+        // TODO: Move to resources
+        let map = Map::new(uvec2(16, 16), "assets/maps/map_01.json").await;
         GameState::add_node(map);
 
         Camera::add_node(vec2(100.0, 100.0));
